Update cachematrix.R
This class is advertised as "for beginners". How infuriatingly misleading. It certainly could be, if it were actually taught that way, rather than consisting of lectures filled with details that are strangely both incomprehensible to new learners and yet, I suspect, redundantly obvious to anyone with sufficient knowledge of R to actually understand them.